replace size_t with uint32_t to avoid wornnings on platforms with different sizes and fix printf  warrnings

+++ src/common/byte_array.c
@@ -24,8 +24,8 @@ struct byte_array NULL_ARRAY = {
 	.ptr = NULL,
 };
 
-
-enum err byte_array_cpy(struct byte_array *dest, const struct byte_array *src, const size_t dest_max_len)
+enum err byte_array_cpy(struct byte_array *dest, const struct byte_array *src,
+			const uint32_t dest_max_len)
 {
 	TRY(_memcpy_s(dest->ptr, dest_max_len, src->ptr, src->len));
 	dest->len = src->len;
